refactor: move invitation jsx and business logic to new component and custom hook resp.
